refactor: task runner
