+++ lib/tools/task_runner/deploy.dart
@@ -1,11 +1,11 @@
 import 'args.dart';
-import 'deployment_runner.dart';
 import 'task.dart';
+import 'task_runner.dart';
 import 'tasks/android_tasks.dart';
-import 'tasks/test.dart';
+import 'tasks/test_task.dart';
 
 void main(List<String> args) {
-  DeploymentRunner(
+  TaskRunner(
     options: ArgOptions.fromArgs(args),
     tasks: [
       LogTask((o) {
@@ -16,8 +16,9 @@ void main(List<String> args) {
             o.mapLabels.keys.where((k) => o.mapLabels[k] == true).toList();
         final msg = [
           [lst.getRange(0, lst.length - 1).join(', '), lst.last]
-              .where((i) => i != null && i.isNotEmpty)
-              .join(' & '),
+                  .where((i) => i != null && i.isNotEmpty)
+                  .join(' & ') +
+              (lst.length == 1 ? ' Only' : ''),
           'Version: ${o.version}',
         ];
         return msg.join('\n');
